fix: subscription to idea db for getting new postcards in real time
