+++ src/app.tsx
@@ -1,4 +1,4 @@
-import { useMemo, useSyncExternalStore } from "react";
+import { useSyncExternalStore } from "react";
 import ReactFlow, { Controls } from "reactflow";
 import "reactflow/dist/style.css";
 import { Postcard } from "./components/postcard";
@@ -6,29 +6,8 @@ import { dbStore } from "./store/store";
 
 const nodeTypes = { postcard: Postcard };
 
-const getRandomCircularCoordinates = () => {
-	// https://stackoverflow.com/questions/5837572/generate-a-random-point-within-a-circle-uniformly
-	const t = 2 * Math.PI * Math.random();
-	const r = Math.sqrt(Math.random());
-	const x = 0.5 + r * Math.cos(t) * 0.75;
-	const y = 0.5 + r * Math.sin(t) * 0.75;
-
-	return { x: x * 1500, y: y * 1000 };
-};
-
 export default function App() {
-	const ideas = useSyncExternalStore(dbStore.subscribe, dbStore.getSnapshot);
-
-	const nodes = useMemo(
-		() =>
-			ideas.map((row) => ({
-				id: row.id,
-				position: getRandomCircularCoordinates(),
-				data: row,
-				type: "postcard",
-			})),
-		[ideas],
-	);
+	const nodes = useSyncExternalStore(dbStore.subscribe, dbStore.getSnapshot);
 
 	return (
 		<div className="h-screen w-screen bg-background">
